Uploaded Data and Scripts (Python Code Files)
Uploaded Our World in Data (OWID) COVID-19 dataset from Oxford University that contains 26 countries that were used for analysis in my research paper. This dataset is in a csv file format. In addition to that, all of the Python Codes that were created for this research in order to analyze the data have been uploaded to my repository.